+++ src/lib.rs
@@ -2,7 +2,8 @@ use crate::fmt_md::MdOptions;
 use crate::fmt_md_inlines::MdInlinesWriterOptions;
 use crate::fmt_plain::PlainOutputOpts;
 use crate::output::{OutputOpts, Stream};
-use crate::select::{ParseError, SelectorAdapter};
+use crate::query::ParseError;
+use crate::select::SelectorAdapter;
 use crate::tree::{InvalidMd, MdDoc, ReadOptions};
 use crate::tree_ref::MdElemRef;
 use crate::tree_ref_serde::SerdeDoc;
