+++ src/query/mod.rs
@@ -1,10 +1,16 @@
 mod query;
 mod selectors;
 
+mod error;
+mod matcher;
+mod matcher_try_from;
+mod selector_try_from;
 mod strings;
 mod traversal;
 mod traversal_composites;
 
+pub use error::*;
+pub use matcher::*;
 #[cfg(test)]
 pub use query::StringVariant;
 #[allow(unused_imports)]
